apply fixes after move
